% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_beds_in_bedset,BEDbase-method}
\alias{bb_beds_in_bedset,BEDbase-method}
\title{Get BEDs associated with BEDset}
\usage{
\S4method{bb_beds_in_bedset}{BEDbase}(x, rec_id, limit = 100, offset = 0, full = FALSE)
}
\arguments{
\item{rec_id}{integer() BEDset record identifier}

\item{limit}{integer() (defaults to 100) maximum records to include}

\item{offset}{integer() (defaults to 0) page offset}

\item{full}{logical() (defaults to FALSE) include all metadata}
}
\value{
list() BED record identifiers
}
\description{
Get BEDs associated with BEDset
}
\examples{
client <- BEDbase()
rec_id <- "gse218680"
bb_beds_in_bedset(client, rec_id)
}
