% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_beds_in_bedset}
\alias{bb_beds_in_bedset}
\title{Get BEDs associated with BEDset}
\usage{
bb_beds_in_bedset(api, bedset_id)
}
\arguments{
\item{api}{API object of BEDbase created from BEDbase()}

\item{bedset_id}{integer() BEDset record identifier}
}
\value{
list() BED record identifiers
}
\description{
Get BEDs associated with BEDset
}
\examples{
api <- BEDbase()
ex_bedset <- bb_example(api, "bedset")
bb_beds_in_bedset(api, ex_bedset$id)

}
