% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_count,BEDbase-method}
\alias{bb_count,BEDbase-method}
\title{Count BEDs or BEDsets}
\usage{
\S4method{bb_count}{BEDbase}(x, rec_type = c("bed", "bedset"))
}
\arguments{
\item{rec_type}{character() bed or bedset}
}
\value{
integer() the number of BEDs or BEDsets available
}
\description{
Count BEDs or BEDsets
}
\examples{
client <- BEDbase()
bb_count(client, "bed")

}
