% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_example,BEDbase-method}
\alias{bb_example,BEDbase-method}
\title{Get example BED or BEDset with metadata}
\usage{
\S4method{bb_example}{BEDbase}(x, rec_type = c("bed", "bedset"))
}
\arguments{
\item{rec_type}{character() bed or bedset}
}
\value{
list() bed or bedset
}
\description{
Get example BED or BEDset with metadata
}
\examples{
client <- BEDbase()
ex_bed <- bb_example(client, "bed")
ex_bedset <- bb_example(client, "bedset")

}
