% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_example}
\alias{bb_example}
\title{Get the example BED file or BEDset with metadata}
\usage{
bb_example(api, rec_type = c("bed", "bedset"))
}
\arguments{
\item{api}{API object of BEDbase created from BEDbase()}

\item{rec_type}{character() bed or bedset}
}
\value{
list() bed or bedset
}
\description{
Get the example BED file or BEDset with metadata
}
\examples{
api <- BEDbase()
bb_example(api, "bed")
bb_example(api, "bedset")

}
