% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_genomes,BEDbase-method}
\alias{bb_genomes,BEDbase-method}
\title{Get genome assemblies in BEDbase}
\usage{
\S4method{bb_genomes}{BEDbase}(x, rec_type = c("bed", "bedset"))
}
\arguments{
\item{rec_type}{character() BED or BEDset}
}
\value{
tibble
}
\description{
Get genome assemblies in BEDbase
}
\examples{
client <- BEDbase()
bb_genomes(client, "bed")

}
