% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_list_beds,BEDbase-method}
\alias{bb_list_beds,BEDbase-method}
\title{List BEDs}
\usage{
\S4method{bb_list_beds}{BEDbase}(x, genome = NULL, bed_type = NULL, limit = 1000, offset = 0)
}
\arguments{
\item{genome}{character() (defaults to NULL) genome keyword}

\item{bed_type}{character() (defaults to NULL) bed file type}

\item{limit}{integer() (defaults to 1000) maximum records}

\item{offset}{integer() (defaults to 0) page token of records}
}
\value{
a tibble of BED records
}
\description{
List BEDs
}
\examples{
client <- BEDbase()
bb_list_beds(client)

}
