% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_list_bedsets,BEDbase-method}
\alias{bb_list_bedsets,BEDbase-method}
\title{List BEDsets}
\usage{
\S4method{bb_list_bedsets}{BEDbase}(x, query = NULL, limit = 1000, offset = 0)
}
\arguments{
\item{query}{character() (defaults to NULL) BEDset keyword}

\item{limit}{integer() (defaults to 1000) maximum records}

\item{offset}{integer() (defaults to 0) page token of records}
}
\value{
a tibble of BEDset records
}
\description{
List BEDsets
}
\examples{
client <- BEDbase()
bb_list_bedsets(client)

}
