% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_metadata,BEDbase-method}
\alias{bb_metadata,BEDbase-method}
\title{Get metadata for an BED, BEDset, or object}
\usage{
\S4method{bb_metadata}{BEDbase}(x, id, rec_type = c("bed", "bedset", "object"))
}
\arguments{
\item{id}{integer() record or object identifier}

\item{rec_type}{character() BED, BEDset, or object}
}
\value{
list() metadata
}
\description{
Get metadata for an BED, BEDset, or object
}
\examples{
client <- BEDbase()
bb_metadata(client, "bed")

}
