% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_metadata_classification,BEDbase-method}
\alias{bb_metadata_classification,BEDbase-method}
\title{Get BED metadata classification}
\usage{
\S4method{bb_metadata_classification}{BEDbase}(x, rec_id)
}
\arguments{
\item{rec_id}{BEDbase record identifier}
}
\value{
tibble() records
}
\description{
Get BED metadata classification
}
\examples{
client <- BEDbase()
bb_metadata_classification(client, "bbad85f21962bb8d972444f7f9a3a932")

}
