% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_metadata_files,BEDbase-method}
\alias{bb_metadata_files,BEDbase-method}
\title{BED files metadata}
\usage{
\S4method{bb_metadata_files}{BEDbase}(x, id)
}
\arguments{
\item{id}{integer() BED record identifier}
}
\value{
tibble() file metadata
}
\description{
BED files metadata
}
\examples{
client <- BEDbase()
ex_bed <- bb_example(client, "bed")
mdf <- bb_metadata_files(client, ex_bed$id)

}
