% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_search,BEDbase-method}
\alias{bb_search,BEDbase-method}
\title{Search BEDbase}
\usage{
\S4method{bb_search}{BEDbase}(x, query, limit = 10, offset = 0)
}
\arguments{
\item{query}{character() keywords to search}

\item{limit}{integer() (defaults to 10) maximum number of results}

\item{offset}{integer() (defaults to 0) page offset of results}
}
\value{
tibble()
}
\description{
Search BEDbase
}
\examples{
client <- BEDbase()
bb_search(client, "excluderanges")

}
