% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_search}
\alias{bb_search}
\title{Search BEDbase}
\usage{
bb_search(x, query, limit = NULL, offset = NULL)
}
\arguments{
\item{query}{character() keywords to search}

\item{limit}{integer() (defaults to NULL) maximum number of results}

\item{offset}{integer() (defaults to NULL) page offset of results}
}
\value{
tibble()
}
\description{
Search BEDbase
}
\examples{
bb_bedbase("excluderanges")

}
