% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_to_granges,BEDbase-method}
\alias{bb_to_granges,BEDbase-method}
\title{Create a GRanges object given a BED id}
\usage{
\S4method{bb_to_granges}{BEDbase}(
  x,
  rec_id,
  file_type = c("bed", "bigbed"),
  extra_cols = c(),
  quietly = FALSE
)
}
\arguments{
\item{rec_id}{BEDbase record identifier}

\item{file_type}{character() bed or bigbed}

\item{extra_cols}{character() extra column names to construct GRanges objects}

\item{quietly}{logical() (defaults to FALSE) display messages}
}
\value{
a GRanges object
}
\description{
Create a GRanges object given a BED id
}
\examples{
client <- BEDbase()
bb_to_granges(client, "421d2128e183424fcc6a74269bae7934")

}
