% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_to_grangeslist,BEDbase-method}
\alias{bb_to_grangeslist,BEDbase-method}
\title{Create a GRangesList object given a BEDset id}
\usage{
\S4method{bb_to_grangeslist}{BEDbase}(x, id, quietly = FALSE)
}
\arguments{
\item{id}{integer() BEDset record identifier}

\item{quietly}{logical() (defaults to FALSE) display messages}
}
\value{
GRangesList() object
}
\description{
Create a GRangesList object given a BEDset id
}
\examples{
client <- BEDbase()
id <- "lola_hg38_ucsc_features"
bb_to_grangeslist(client, id)

}
