% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_to_grangeslist}
\alias{bb_to_grangeslist}
\title{Create a GRangesList object given a BEDset id}
\usage{
bb_to_grangeslist(bedbase, bedset_id, quietly = TRUE)
}
\arguments{
\item{bedbase}{BEDbase() object}

\item{bedset_id}{integer(1) BEDset record identifier}

\item{quietly}{logical(1) (default \code{TRUE}) display messages}
}
\value{
\link[GenomicRanges:GRangesList-class]{GRangesList}
}
\description{
Create a GRangesList object given a BEDset id
}
\examples{
bedbase <- BEDbase()
bedset_id <- "lola_hg38_ucsc_features"
bb_to_grangeslist(bedbase, bedset_id)

}
