% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bedbaser}
\alias{bedbaser}
\title{An R client for BEDbase}
\usage{
bedbaser()
}
\value{
BEDbase object
}
\description{
bedbaser exposes the BEDbase API and includes convenience
functions for common tasks, including importing a BED id into \code{GRanges}
object and a BEDset id into a \code{GRangesList}.
}
\details{
The main functions are as follows
\itemize{
\item \code{bedbaser::bedbaser()}: API constructor
\item \code{bedbaser::bb_example()}: Retrieve an example BED file or BEDset
\item \code{bedbaser::bb_metadata()}: Retrieve metadata for a BED file or BEDset
\item \code{bedbaser::bb_list_beds()}: List all BED files
\item \code{bedbaser::bb_list_bedsets()}: List all BEDsets
\item \code{bedbaser::bb_beds_in_bedset()}: List BED files in BEDset
\item \code{bedbaser::bb_bed_text_search()}: Search BED files by text
\item \code{bedbaser::bb_to_granges()}: Create a \code{GRanges} object from a BED id
\item \code{bedbaser::bb_to_grangeslist()}: Create a \code{GrangesList} from a BEDset id
Set the environment variable \code{BEDBASER_CACHE} to alter the cache path for
downloaded BED files.
}
}
\examples{
bedbaser()

}
