% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{count}
\alias{count}
\title{Count BEDs or BEDsets}
\usage{
count(rec_type = c("bed", "bedset"), quiet = FALSE)
}
\arguments{
\item{rec_type}{character() bed or bedset}

\item{quiet}{logical() (default FALSE) display message}
}
\value{
integer() the number of BEDs or BEDsets available
}
\description{
Count BEDs or BEDsets
}
\examples{
count("bed")

}
