% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.bed_file_to_granges}
\alias{.bed_file_to_granges}
\title{Create GRanges object from a BED file}
\usage{
.bed_file_to_granges(file_path, metadata, extra_cols = NULL, quietly = FALSE)
}
\arguments{
\item{file_path}{character() path to BED file}

\item{metadata}{list() full metadata}

\item{extra_cols}{character() (defaults to NULL) extra column names to
construct GRanges objects}

\item{quietly}{boolean() (default FALSE) Display information messages}
}
\value{
GRanges() object representing BED
}
\description{
If the BED format is known, \code{extra_cols} may be used to set the column name
and type. For example, \code{extra_cols = c(signalValue = "numeric", pValue = "numeric", qValue = "numeric")}.
}
\details{
It will also attempt to supply the \code{genome} if \code{genome_alias} exists in the
metadata.

Aborts if the length of \code{extra_cols} is not equal to Y in BEDX+Y.
}
\examples{
client <- BEDbase()
ex_bed <- bb_example(client, "bed")
md <- bb_metadata(client, ex_bed$id, "bed", TRUE)
file_path <- .get_file(md, "bed", "http")
.bed_file_to_granges(file_path, md)
}
