% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.BEDBaseBaseUrl}
\alias{.BEDBaseBaseUrl}
\title{Base URL of BEDbase}
\format{
An object of class \code{character} of length 1.
}
\usage{
.BEDBaseBaseUrl
}
\description{
Base URL of BEDbase
}
\keyword{datasets}
