% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.download_and_cache}
\alias{.download_and_cache}
\title{Retrieve path from cache or download file and cache}
\usage{
.download_and_cache(url, quiet = FALSE)
}
\arguments{
\item{url}{character() remote resource}

\item{quiet}{logical() display message}
}
\value{
filepath character()
}
\description{
Note: This function is described in BiocFileCache
}
\examples{
url <- "https://data2.bedbase.org/bed_files/hg38_UCSC_telomere.bed.gz"
filepath <- .download_and_cache(url)
}
