% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.file_to_granges}
\alias{.file_to_granges}
\title{Format BED file metadata}
\usage{
.file_to_granges(file_path, bed_type, bed_format)
}
\arguments{
\item{file_path}{character() path to BED file}

\item{bed_type}{character() bed type}

\item{bed_format}{character() format name}
}
\value{
GRanges()
}
\description{
Format BED file metadata
}
\examples{
.file_to_granges(file_path, bed_type, bed_format)
}
