% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_metadata_files}
\alias{.format_metadata_files}
\title{Format BED file metadata}
\usage{
.format_metadata_files(metadata)
}
\arguments{
\item{records}{list() metadata}
}
\value{
tibble() file metadata
}
\description{
Format BED file metadata
}
\examples{
client <- BEDbase()
ex_bed <- bb_example(client, "bed")
ex_metadata <-bb_metadata(client, ex_bed$id, "bed", TRUE)
.format_metadata_files(ex_bed$files)
}
