% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_extra_cols}
\alias{.get_extra_cols}
\title{Return a named vector with type}
\usage{
.get_extra_cols(file_path, x, y)
}
\arguments{
\item{file_path}{character() path to BED}

\item{x}{double() the x in BEDX+Y}

\item{y}{double() the y in BEDX+Y}
}
\value{
vector representing extraCols for rtracklayer
}
\description{
Return a named vector with type
}
\examples{
id <- "608827efc82fcaa4b0bfc65f590ffef8"
md <- bb_metadata(client, id, "bed", TRUE)
file_path <- .get_file_path(md$files$bed_file$access_methods[[1]]$access_url$url,
                            "bed")
.get_extra_cols(file_path, 3, 9)
}
