% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_extraCols}
\alias{.get_extraCols}
\title{Return a named vector with type}
\usage{
.get_extraCols(file_path, y)
}
\arguments{
\item{file_path}{character() path to BED}

\item{y}{double() the y in BEDX+Y}
}
\value{
vector representing extraCols for rtracklayer
}
\description{
Return a named vector with type
}
\examples{

.get_extraCols("path/to/my/file.bed", 3)
}
