% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_file}
\alias{.get_file}
\title{Get a file from BEDbase}
\usage{
.get_file(
  metadata,
  file_type = c("bed", "bigbed"),
  access_type = c("s3", "http"),
  quietly = FALSE
)
}
\arguments{
\item{metadata}{list() full metadata}

\item{file_type}{character() bed or bigbed}

\item{access_type}{character() s3 or http}

\item{quietly}{logical() (defaults to FALSE) display messages}
}
\value{
character() file path
}
\description{
Get a file from BEDbase
}
\examples{
client <- BEDbase()
ex_bed <- bb_example(client, "bed")
md <- bb_metadata(client, ex_bed$id, "bed", TRUE)
.get_file(md, "bed", "http")
}
