% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_file_path}
\alias{.get_file_path}
\title{Get a file from BEDbase}
\usage{
.get_file_path(file_url, file_type, quietly = FALSE)
}
\arguments{
\item{file_url}{character() url to file}

\item{file_type}{character() bed or bigbed}

\item{quietly}{logical() (defaults to FALSE) display messages}
}
\value{
character() file path
}
\description{
Get a file from BEDbase
}
\examples{
client <- BEDbase()
ex_bed <- bb_example(client, "bed")
md <- bb_metadata(client, ex_bed$id, "bed", TRUE)
.get_file_path(md$files$bed_file$access_methods[[1]]$access_url$url, "bed")
}
