% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_format}
\alias{.get_format}
\title{Get format}
\usage{
.get_format(file_path, data_format)
}
\arguments{
\item{file_path}{character(1) path to BED file}

\item{data_format}{character(1) bed file format}
}
\value{
character(1) format
}
\description{
Get format for \link[rtracklayer:BEDFile-class]{rtracklayer}
For supported formats, see
\url{https://docs.bedbase.org/bedbase/user/bed_classification/}.
}
\examples{
bedbase <- BEDbase()
ex_bed <- bb_example(bedbase, "bed")
md <- bb_metadata(bedbase, ex_bed$id, TRUE)
file_path <- .get_file(md$bed_compliance, getCache(bedbase), "bed", "http")
format <- .get_format(md$data_format, file_path)
}
