% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getAccessIds}
\alias{.getAccessIds}
\title{Get valid access ids}
\usage{
.getAccessIds(object_id)
}
\arguments{
\item{object_id}{BEDbase object id}
}
\value{
a vector with available access_ids
}
\description{
Note: omits 'local' option
}
\examples{
.getAccessIds("bed.421d2128e183424fcc6a74269bae7934.bedfile")
}
