% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getObjectId}
\alias{.getObjectId}
\title{Construct object identifier}
\usage{
.getObjectId(record_id, record_type = "bed", result_id = "bedfile")
}
\arguments{
\item{record_id}{BEDbase record id}

\item{record_type}{BEDbase record type}

\item{result_id}{BEDbase result id}
}
\value{
String
}
\description{
Construct object identifier
}
\examples{
object_id <- .getObjectId("eaf9ee97241f300f1c7e76e1f945141f")
}
