% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.importToGRanges}
\alias{.importToGRanges}
\title{Download BED file associated with record_id and create a GRanges object}
\usage{
.importToGRanges(record_id)
}
\arguments{
\item{record_id}{BEDbase record id}
}
\description{
Download BED file associated with record_id and create a GRanges object
}
\examples{
grobj <- .importToGRanges("421d2128e183424fcc6a74269bae7934")
}
