% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ungzip}
\alias{.ungzip}
\title{Decompress gunzip file}
\usage{
.ungzip(file_path)
}
\arguments{
\item{file_path}{character() path to gunzip file}
}
\value{
character() decompressed file path
}
\description{
Decompress gunzip file
}
\examples{
.ungzip("path/to/my.gzip")
}
