% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{download_bed_file}
\alias{download_bed_file}
\title{Download BED}
\usage{
download_bed_file(
  record_id,
  destdir,
  file_type = c("bytes", "thumbnail"),
  access_id = "http",
  quiet = TRUE
)
}
\arguments{
\item{record_id}{integer(1) BED record identifier}

\item{destdir}{character(1) directory to save the file}

\item{file_type}{character(1) type of file to download, bytes or thumbnail}

\item{access_id}{character(1) (default http) download protocol, "local"
access_identifier is removed.}

\item{quiet}{logical(1) (default TRUE) if FALSE, prints message}
}
\value{
file path
}
\description{
Download BED
}
\examples{
filepath <-download_bed_file("421d2128e183424fcc6a74269bae7934", "/tmp", "bytes")
filepath <-download_bed_file("ce9f6e8adf5779edb632245a166b92bd", "/tmp", "bytes")
zebrafish <-download_bed_file("eec93c0649bfc2fc90cdc9fca3f52588", "/tmp", "bytes")
filepath <-download_bed_file("e6cd0dbb252be66f8c2bdb0d23ed431c", "/tmp", "bytes")
telomere <-download_bed_file("b56d842d5727f1cfaaccb0bdc1bba6ad", "/tmp", "bytes")

}
