% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{download_file}
\alias{download_file}
\title{Download BED or thumbnail}
\usage{
download_file(
  rec_id,
  file_type = c("bytes", "thumbnail"),
  acc_id = "http",
  quiet = FALSE
)
}
\arguments{
\item{rec_id}{integer() BED record identifier}

\item{file_type}{character() type of file to download, bytes or thumbnail}

\item{acc_id}{character() (default http) access identifier, "local"
is removed.}

\item{quiet}{logical() (default FALSE) display message}
}
\value{
character() path to file
}
\description{
Download BED or thumbnail
}
\examples{
telomere <- download_file("80d2b2581bb25fa6b73ec56c11969fb3", "bytes")

}
