% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDbaseR.R
\name{downloadBedFile}
\alias{downloadBedFile}
\title{Download BED file}
\usage{
downloadBedFile(record_id, destdir, access_id = "http")
}
\arguments{
\item{record_id}{BEDbase record id}

\item{destdir}{directory to save the file}

\item{access_id}{Download protocol. "Local" access_id is removed.}
}
\value{
file path
}
\description{
Download BED file
}
\examples{
filepath <-downloadBedFile("421d2128e183424fcc6a74269bae7934", "/tmp")

}
