% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_access_ids}
\alias{get_access_ids}
\title{Get valid access identifiers}
\usage{
get_access_ids(obj_id, quiet = FALSE)
}
\arguments{
\item{obj_id}{character(1) BEDbase object record identifier}

\item{quiet}{logical(1) (default FALSE) display message}
}
\value{
character() available access identifiers
}
\description{
Note: omits 'local' option
}
\examples{
get_access_ids("bed.421d2128e183424fcc6a74269bae7934.bedfile")
}
