% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_bb_service_info,BEDbase-method}
\alias{get_bb_service_info,BEDbase-method}
\title{Get service information}
\usage{
\S4method{get_bb_service_info}{BEDbase}(x)
}
\value{
list() service info, such as version
}
\description{
Get service information
}
\examples{
bb <- BEDbase()
get_bb_service_info(bb)

}
