% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{get_beds_in_bedset}
\alias{get_beds_in_bedset}
\title{Get BEDs associated with BEDset}
\usage{
get_beds_in_bedset(record_id)
}
\arguments{
\item{record_id}{integer(1) BEDset record identifier}
}
\value{
list() BED record identifiers
}
\description{
Get BEDs associated with BEDset
}
\examples{
get_beds_in_bedset("excluderanges")

}
