% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{get_genomes}
\alias{get_genomes}
\title{Get genome assemblies in BEDbase}
\usage{
get_genomes(rec_type = c("bed", "bedset"), quiet = FALSE)
}
\arguments{
\item{rec_type}{character() BED, BEDset, or object}

\item{quiet}{logical() (default FALSE) display message}
}
\value{
list() metadata
}
\description{
Get genome assemblies in BEDbase
}
\examples{
get_genomes("bed")

}
