% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata for an BED, BEDset, or object}
\usage{
get_metadata(id, rec_type = c("bed", "bedset", "object"))
}
\arguments{
\item{id}{integer(1) record or object identifier}

\item{rec_type}{character(1) BED, BEDset, or object}
}
\value{
list() metadata
}
\description{
Get metadata for an BED, BEDset, or object
}
\examples{
get_metadata("421d2128e183424fcc6a74269bae7934", "bed")

}
