% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rec_type}
\alias{get_rec_type}
\title{Get record type}
\usage{
get_rec_type(id)
}
\arguments{
\item{id}{character() BEDbase identifier}
}
\value{
logical()
}
\description{
Get record type
}
\examples{
get_rec_type("eaf9ee97241f300f1c7e76e1f945141f")

}
