% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{get_records}
\alias{get_records}
\title{Get record identifiers and names for BEDs or BEDsets}
\usage{
get_records(rec_type = c("bed", "bedset"), quiet = FALSE)
}
\arguments{
\item{rec_type}{character() bed or bedset}

\item{quiet}{logical() (default FALSE) display message}
}
\value{
a tibble of record identifiers and record names
}
\description{
Note: Should be paged
}
\examples{
recs <- get_records("bed")

}
