% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_service_info}
\alias{get_service_info}
\title{Get service information}
\usage{
get_service_info()
}
\value{
list() service info, such as version
}
\description{
Get service information
}
\examples{
get_service_info()
}
