% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDbaseR.R
\name{getBedsIn}
\alias{getBedsIn}
\title{Get BED record identifiers associated with BEDset}
\usage{
getBedsIn(record_id)
}
\arguments{
\item{record_id}{record identifier}
}
\value{
a list of record identifiers
}
\description{
Get BED record identifiers associated with BEDset
}
\examples{
getBedsIn("excluderanges")

}
