% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{listRecords}
\alias{listRecords}
\title{List BED or BEDsets}
\usage{
listRecords(record_type = c("bed", "bedset"))
}
\arguments{
\item{record_type}{bed or bedset}
}
\value{
a list of bed or bedsets

a list with metadata
}
\description{
List BED or BEDsets
}
\examples{
recs <- listRecords("bed")

}
