% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{query_bedbase}
\alias{query_bedbase}
\title{Query BEDbase API}
\usage{
query_bedbase(endpoint, quiet = FALSE)
}
\arguments{
\item{endpoint}{character(1) BEDbase API endpoint}

\item{quiet}{logical(1) (default FALSE) display message}
}
\value{
a vector or list
}
\description{
Note: The genomes end point error
}
\examples{
query_bedbase("bed/count")
}
