% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{schemas,BEDbase-method}
\alias{schemas,BEDbase-method}
\title{Display schemas}
\usage{
\S4method{schemas}{BEDbase}(x)
}
\arguments{
\item{x}{BEDbase(1) object}
}
\value{
list(1) API end points
}
\description{
Display schemas
}
\examples{
bedbase <- BEDbase()
schemas(bedbase)

}
