% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{search_bedbase}
\alias{search_bedbase}
\title{Search BEDbase}
\usage{
search_bedbase(keywords, limit, quiet = FALSE)
}
\arguments{
\item{keywords}{character() keyword to search for}

\item{limit}{integer() maximum number of results to include}

\item{quiet}{logical() (default FALSE) display message}
}
\value{
tibble()
}
\description{
Search BEDbase
}
\examples{
search_bedbase("excluderanges")

}
