% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{setCache,BEDbase-method}
\alias{setCache,BEDbase-method}
\title{Set cache path}
\usage{
\S4method{setCache}{BEDbase}(x, cache_path, quietly = TRUE)
}
\arguments{
\item{x}{BEDbase object}

\item{cache_path}{character()}

\item{quietly}{(default TRUE) display messages}
}
\value{
BiocFileCache cache of BED files
}
\description{
Set cache path
}
\examples{
api <- BEDbase(tempdir())
api <- setCache(api, "/tmp")

}
