% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{tags,BEDbase-method}
\alias{tags,BEDbase-method}
\title{Display functions for a tag}
\usage{
\S4method{tags}{BEDbase}(x, .tags, .deprecated = FALSE)
}
\arguments{
\item{x}{BEDbase() object}

\item{.tags}{character() tags for filtering operations}

\item{.deprecated}{(default \code{FALSE}) if deprecated}
}
\value{
list() API endpoints
}
\description{
Display functions available through the API associated with a
tag keyword in bedhost.
}
\examples{
bedbase <- BEDbase()
unique(tags(bedbase)$tag)
tags(bedbase, "bedset")

}
