% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_bed_text_search}
\alias{bb_bed_text_search}
\title{Search BED files by text}
\usage{
bb_bed_text_search(api, query, limit = 10, offset = 0)
}
\arguments{
\item{api}{API object of BEDbase created from BEDbase()}

\item{query}{character() keywords to search}

\item{limit}{integer() (default 10) maximum number of results}

\item{offset}{integer() (default 0) page offset of results}
}
\value{
tibble()
}
\description{
Returns all available results scored.
}
\examples{
api <- BEDbase()
bb_bed_text_search(api, "hg38")

}
