% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_beds_in_bedset,BEDbase-method}
\alias{bb_beds_in_bedset,BEDbase-method}
\title{Get BEDs associated with BEDset}
\usage{
\S4method{bb_beds_in_bedset}{BEDbase}(x, id)
}
\arguments{
\item{id}{integer() BEDset record identifier}
}
\value{
list() BED record identifiers
}
\description{
Get BEDs associated with BEDset
}
\examples{
client <- BEDbase()
ex_bedset <- bb_example(client, "bedset")
bb_beds_in_bedset(client, ex_bedset$id)

}
