% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_beds_in_bedset,BEDbase-method}
\alias{bb_beds_in_bedset,BEDbase-method}
\title{Get BEDs associated with BEDset}
\usage{
\S4method{bb_beds_in_bedset}{BEDbase}(x, rec_id)
}
\arguments{
\item{rec_id}{integer() BEDset record identifier}
}
\value{
list() BED record identifiers
}
\description{
Get BEDs associated with BEDset
}
\examples{
client <- BEDbase()
bb_beds_in_bedset(client, "bed")
rec_id <- "421d2128e183424fcc6a74269bae7934"
}
