% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_beds_in_bedset}
\alias{bb_beds_in_bedset}
\title{Get BEDs associated with BEDset}
\usage{
bb_beds_in_bedset(bedbase, bedset_id)
}
\arguments{
\item{bedbase}{API object of BEDbase created from BEDbase()}

\item{bedset_id}{integer(1) BEDset record identifier}
}
\value{
tibble(1) information of BED files in BEDset
}
\description{
Get BEDs associated with BEDset
}
\examples{
bedbase <- BEDbase()
ex_bedset <- bb_example(bedbase, "bedset")
bb_beds_in_bedset(bedbase, ex_bedset$id)

}
