% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_example}
\alias{bb_example}
\title{Get the example BED file or BEDset with metadata}
\usage{
bb_example(bedbase, rec_type = c("bed", "bedset"))
}
\arguments{
\item{bedbase}{API object of BEDbase created from BEDbase()}

\item{rec_type}{character(1) bed or bedset}
}
\value{
list(1) bed files or bedsets
}
\description{
Get the example BED file or BEDset with metadata
}
\examples{
bedbase <- BEDbase()
bb_example(bedbase, "bed")
bb_example(bedbase, "bedset")

}
