% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_list_beds}
\alias{bb_list_beds}
\title{List BEDs}
\usage{
bb_list_beds(api, genome = NULL, bed_type = NULL, limit = 1000, offset = 0)
}
\arguments{
\item{api}{API object of BEDbase created from bedbaser()}

\item{genome}{character() (default NULL) genome keyword}

\item{bed_type}{character() (default NULL) bed file type}

\item{limit}{integer() (default 1000) maximum records}

\item{offset}{integer() (default 0) page token of records}
}
\value{
tibble() of BED records
}
\description{
List BEDs
}
\examples{
api <- bedbaser()
bb_list_beds(api)

}
