% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_list_bedsets}
\alias{bb_list_bedsets}
\title{List BEDsets}
\usage{
bb_list_bedsets(api, query = NULL, limit = 1000, offset = 0)
}
\arguments{
\item{api}{API object of BEDbase created from bedbaser()}

\item{query}{character() (default NULL) keyword}

\item{limit}{integer() (default 1000) maximum records of bedsets}

\item{offset}{integer() (default 0) page token of records}
}
\value{
tibble() of BEDset records
}
\description{
List BEDsets
}
\examples{
api <- bedbaser()
bb_list_bedsets(api)

}
