% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_metadata,BEDbase-method}
\alias{bb_metadata,BEDbase-method}
\title{Get metadata for an BED or BEDset}
\usage{
\S4method{bb_metadata}{BEDbase}(x, id, rec_type = c("bed", "bedset"), full = FALSE)
}
\arguments{
\item{id}{integer() record or object identifier}

\item{rec_type}{character() BED or BEDset}

\item{full}{logical() (default FALSE) include full record with stats, files,
and metadata}
}
\value{
list() metadata
}
\description{
Aborts if not found or id is not not 32 characters
}
\examples{
client <- BEDbase()

ex_bed <- bb_example(client, "bed")
md <- bb_metadata(client, ex_bed$id, "bed")

ex_bedset <- bb_example(client, "bedset")
md <- bb_metadata(client, ex_bedset$id, "bedset")

}
