% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_metadata}
\alias{bb_metadata}
\title{Get metadata for a BED file or BEDset}
\usage{
bb_metadata(api, id, full = FALSE)
}
\arguments{
\item{api}{API object of BEDbase created from BEDbase()}

\item{id}{integer() record or object identifier}

\item{full}{logical() (default FALSE) include full record with stats, files,
and metadata}
}
\value{
list() metadata
}
\description{
Get metadata for a BED file or BEDset. Abort
if not found or id is not not 32 characters
}
\examples{
api <- BEDbase()

ex_bed <- bb_example(api, "bed")
bb_metadata(api, ex_bed$id)

ex_bedset <- bb_example(api, "bedset")
bb_metadata(api, ex_bedset$id)

}
