% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_metadata_files,BEDbase-method}
\alias{bb_metadata_files,BEDbase-method}
\title{BED files metadata}
\usage{
\S4method{bb_metadata_files}{BEDbase}(x, rec_id)
}
\arguments{
\item{rec_id}{integer() BED record identifier}
}
\value{
tibble() records
}
\description{
BED files metadata
}
\examples{
client <- BEDbase()
metadata <- bb_metadata_files(client, "bbad85f21962bb8d972444f7f9a3a932")

}
