% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_records,BEDbase-method}
\alias{bb_records,BEDbase-method}
\title{Get record identifiers and names for BEDs or BEDsets}
\usage{
\S4method{bb_records}{BEDbase}(x, rec_type = c("bed", "bedset"), limit = NULL, token = NULL)
}
\arguments{
\item{rec_type}{character() bed or bedset}

\item{limit}{integer() (defaults to NULL) maximum records}

\item{token}{integer() (defaults to NULL) page token of records}
}
\value{
a tibble of record identifiers and record names
}
\description{
Note: how to get next page token
}
\examples{
client <- BEDbase()
bb_records(client, "bed")

}
