% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_save}
\alias{bb_save}
\title{Save a BED or BEDset files to a path given an id}
\usage{
bb_save(
  api,
  bed_or_bedset_id,
  path,
  file_type = "bed",
  access_type = "http",
  quietly = TRUE
)
}
\arguments{
\item{api}{API object of BEDbase created from BEDbase()}

\item{bed_or_bedset_id}{integer(1) BED or BEDset record identifier}

\item{path}{character(1) directory to save file}

\item{file_type}{character(1) (default \code{"bed"}) bed, bigbed, etc.}

\item{access_type}{character(1) (default \code{"http"})  s3 or http}

\item{quietly}{logical(1) (default \code{TRUE}) display messages}
}
\value{
An invisible \code{NULL}
}
\description{
Save a BED or BEDset files to a path given an id
}
\examples{
api <- BEDbase()
ex_bed <- bb_example(api, "bed")
bb_save(api, ex_bed$id, tempdir())

}
