% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_save}
\alias{bb_save}
\title{Save a BED or BEDset files to a path given an id}
\usage{
bb_save(
  api,
  bed_or_bedset_id,
  path,
  file_type = "bed",
  access_type = "http",
  quietly = TRUE
)
}
\arguments{
\item{api}{API object of BEDbase created from BEDbase()}

\item{bed_or_bedset_id}{integer() BED or BEDset record identifier}

\item{path}{character() directory to save file}

\item{file_type}{character() (default bed) bed, bigbed, etc.}

\item{access_type}{character() (default http)  s3 or http}

\item{quietly}{logical() (default TRUE) display messages}
}
\description{
Save a BED or BEDset files to a path given an id
}
\examples{
api <- BEDbase()
ex <- bb_example(api, "bed") 
bb_save(api, ex$id, tempdir())

}
