% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_stats}
\alias{bb_stats}
\title{Get BEDbase statistics}
\usage{
bb_stats(bedbase, detailed = FALSE)
}
\arguments{
\item{bedbase}{BEDbase() object}

\item{detailed}{logical(1) if TRUE display detailed information}
}
\value{
An invisible \code{NULL}
}
\description{
Get statistics on available BED files, BEDsets, and genomoes.
}
\examples{
bedbase <- BEDbase()
bb_stats(bedbase)
}
