% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_to_granges,BEDbase-method}
\alias{bb_to_granges,BEDbase-method}
\title{Create a GRanges object given a BED id}
\usage{
\S4method{bb_to_granges}{BEDbase}(
  x,
  bed_id,
  file_type = c("bed", "bigbed"),
  extra_cols = NULL,
  quietly = FALSE
)
}
\arguments{
\item{bed_id}{integer() BED record identifier}

\item{file_type}{character() bed or bigbed}

\item{extra_cols}{character() (defaults to NULL) extra column names to
construct GRanges objects}

\item{quietly}{logical() (defaults to FALSE) display messages}
}
\value{
GRanges() object
}
\description{
Create a GRanges object given a BED id
}
\examples{
client <- BEDbase()
ex_bed <- bb_example(client, "bed")
bb_to_granges(client, ex_bed$id)

}
