% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_to_granges}
\alias{bb_to_granges}
\title{Create a GRanges object given a BED id}
\usage{
bb_to_granges(
  bedbase,
  bed_id,
  file_type = "bed",
  extra_cols = NULL,
  quietly = TRUE
)
}
\arguments{
\item{bedbase}{API object of BEDbase created from \code{BEDbase()}}

\item{bed_id}{integer(1) BED record identifier}

\item{file_type}{character(1) bed or bigbed}

\item{extra_cols}{character() (default \code{NULL}) extra column names to
construct GRanges objects}

\item{quietly}{logical(1) (default \code{TRUE}) display messages}
}
\value{
GRanges(1) object
}
\description{
Generates the column and types for broad and narrow peak files.
}
\examples{
bedbase <- BEDbase()
ex_bed <- bb_example(bedbase, "bed")
bb_to_granges(bedbase, ex_bed$id)

}
