% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_to_granges}
\alias{bb_to_granges}
\title{Create a GRanges object given a BED id}
\usage{
bb_to_granges(
  api,
  bed_id,
  file_type = "bed",
  extra_cols = NULL,
  quietly = TRUE
)
}
\arguments{
\item{api}{API object of BEDbase created from bedbaser()}

\item{bed_id}{integer() BED record identifier}

\item{file_type}{character() bed or bigbed}

\item{extra_cols}{character() (default NULL) extra column names to
construct GRanges objects}

\item{quietly}{logical() (default TRUE) display messages}
}
\value{
GRanges() object
}
\description{
Generates the column and types for broad and narrow peak files.
}
\examples{
api <- bedbaser()
ex_bed <- bb_example(api, "bed")
bb_to_granges(api, ex_bed$id)

}
