% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{bb_to_grangeslist}
\alias{bb_to_grangeslist}
\title{Create a GRangesList object given a BEDset id}
\usage{
bb_to_grangeslist(api, bedset_id, quietly = TRUE)
}
\arguments{
\item{api}{API object of BEDbase created from BEDbase()}

\item{bedset_id}{integer() BEDset record identifier}

\item{quietly}{logical() (default TRUE) display messages}
}
\value{
GRangesList() object
}
\description{
Create a GRangesList object given a BEDset id
}
\examples{
api <- BEDbase()
id <- "lola_hg38_ucsc_features"
bb_to_grangeslist(api, id)

}
