% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{BEDbase}
\alias{BEDbase}
\alias{getCache}
\alias{setCache}
\title{An R client for BEDbase}
\usage{
BEDbase(cache_path)

getCache(x, quietly = TRUE)

setCache(x, cache_path, quietly = TRUE)
}
\arguments{
\item{cache_path}{character()}

\item{x}{BEDbase object}

\item{quietly}{(default TRUE) display messages}
}
\value{
BEDbase object
}
\description{
bedbaser exposes the BEDbase API and includes convenience
functions for common tasks, such as to import a BED file by id into a
\code{GRanges} object and a BEDset by its id into a \code{GRangesList}.
}
\details{
The convenience functions are as follows
\itemize{
\item \code{bedbaser::BEDbase()}: API service constructor
\item \code{bedbaser::getCache()}: Retrieve cache
\item \code{bedbaser::setCache()}: Set path to cache
\item \code{bedbaser::bb_example()}: Retrieve an example BED file or BEDset
\item \code{bedbaser::bb_metadata()}: Retrieve metadata for a BED file or BEDset
\item \code{bedbaser::bb_list_beds()}: List all BED files
\item \code{bedbaser::bb_list_bedsets()}: List all BEDsets
\item \code{bedbaser::bb_beds_in_bedset()}: List BED files in BEDset
\item \code{bedbaser::bb_bed_text_search()}: Search BED files by text
\item \code{bedbaser::bb_to_granges()}: Create a \code{GRanges} object from a BED id
\item \code{bedbaser::bb_to_grangeslist()}: Create a \code{GrangesList} from a BEDset id
}
}
\examples{
BEDbase()

}
