% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser-package.R
\docType{package}
\name{bedbaser-package}
\alias{bedbaser-package}
\title{bedbaser: A BEDbase client}
\description{
A client for BEDbase. bedbaser provides access to the API at api.bedbase.org. It also includes convenience functions to import BED files into GRanges objects and BEDsets into GRangesLists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/waldronlab/bedbaser}
  \item Report bugs at \url{https://github.com/jwokaty/bedbaser/issues}
}

}
\author{
\strong{Maintainer}: Jen Wokaty \email{jennifer.wokaty@sph.cuny.edu} (\href{https://orcid.org/0009-0008-0900-8793}{ORCID})

Authors:
\itemize{
  \item Levi Waldron \email{levi.waldron@sph.cuny.edu} (\href{https://orcid.org/0000-0003-2725-0694}{ORCID})
}

}
\keyword{internal}
