% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{count}
\alias{count}
\title{Count BEDs or BEDsets}
\usage{
count(record_type = c("bed", "bedset"))
}
\arguments{
\item{record_type}{character(1) bed or bedset}
}
\value{
integer(1) the number of BEDs or BEDsets available
}
\description{
Count BEDs or BEDsets
}
\examples{
count("bed")

}
