% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.file_to_granges}
\alias{.file_to_granges}
\title{Create GRanges object from a BED file}
\usage{
.file_to_granges(file_path, bed_type, bed_format, extra_cols, quietly = FALSE)
}
\arguments{
\item{file_path}{character() path to BED file}

\item{bed_type}{character() bed type}

\item{bed_format}{character() format name}

\item{extra_cols}{character() extra column names to construct GRanges objects}

\item{quietly}{boolean() (default FALSE) Display information messages}
}
\value{
GRanges() object representing BED
}
\description{
Create GRanges object from a BED file
}
\examples{
.file_to_granges(file_path, bed_type, bed_format)
}
