% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_metadata_files}
\alias{.format_metadata_files}
\title{Format BED file metadata}
\usage{
.format_metadata_files(records)
}
\arguments{
\item{records}{list() metadata records}
}
\value{
tibble() file metadata
}
\description{
Format BED file metadata
}
\examples{
client <- BEDbase()
example <- content(client$get_example_bed_record_v1_bed_example_get())
.format_metadata_files(example$files)
}
