% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getAccessIds}
\alias{.getAccessIds}
\title{Get valid access ids}
\usage{
.getAccessIds(object_id)
}
\arguments{
\item{record_id}{BEDbase record id}
}
\value{
a vector available access_ids
}
\description{
Note: omits 'local' option
}
\examples{
object_id <- .getObjectId("421d2128e183424fcc6a74269bae7934")
access_id <- getAccessIds(object_id)
}
