% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDbaseR.R
\name{downloadBedFile}
\alias{downloadBedFile}
\title{Download BED file}
\usage{
downloadBedFile(record_id, destdir, access_id = "http")
}
\arguments{
\item{record_id}{BEDbase record id}

\item{destdir}{directory to save the file}

\item{access_id}{Download protocol. "Local" access_id is removed.}
}
\value{
file path
}
\description{
Download BED file
}
\examples{
filepath <-downloadBedFile("421d2128e183424fcc6a74269bae7934", "/tmp")
filepath <-downloadBedFile("ce9f6e8adf5779edb632245a166b92bd", "/tmp")
zebrafish <-downloadBedFile("eec93c0649bfc2fc90cdc9fca3f52588", "/tmp")
filepath <-downloadBedFile("e6cd0dbb252be66f8c2bdb0d23ed431c", "/tmp")
telomere <-downloadBedFile("b56d842d5727f1cfaaccb0bdc1bba6ad", "/tmp")

}
