% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDbaseR.R
\name{fileToGRanges}
\alias{fileToGRanges}
\title{Create a GRanges object from a BED file}
\usage{
fileToGRanges(filepath)
}
\arguments{
\item{filepath}{path to a BED file}
}
\value{
a GRanges object
}
\description{
Create a GRanges object from a BED file
}
\examples{
grobj <- fileToGRanges(filepath)

}
