% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{get_beds_in_bedset}
\alias{get_beds_in_bedset}
\title{Get BEDs associated with BEDset}
\usage{
get_beds_in_bedset(rec_id, quiet = FALSE)
}
\arguments{
\item{rec_id}{integer() BEDset record identifier}

\item{quiet}{logical() (default FALSE) display message}
}
\value{
list() BED record identifiers
}
\description{
Get BEDs associated with BEDset
}
\examples{
get_beds_in_bedset("excluderanges")

}
