% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{getCache,BEDbase-method}
\alias{getCache,BEDbase-method}
\title{Return cache path}
\usage{
\S4method{getCache}{BEDbase}(x, cache_type = c("bedfiles", "bedsets"))
}
\arguments{
\item{x}{BEDbase() object}

\item{cache_type}{character(1) bedfiles or bedsets}
}
\value{
BiocFileCache() object of BED files
}
\description{
Return cache path
}
\examples{
bedbase <- BEDbase(tempdir())
getCache(bedbase, "bedfiles")

}
