% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{getCache,BEDbase-method}
\alias{getCache,BEDbase-method}
\title{Return cache path}
\usage{
\S4method{getCache}{BEDbase}(x, quietly = TRUE)
}
\arguments{
\item{x}{BEDbase(1) object}

\item{quietly}{logical(1) (default \code{TRUE}) display messages}
}
\value{
BiocFileCache(1) object of BED files
}
\description{
Return cache path
}
\examples{
api <- BEDbase(tempdir())
getCache(api)

}
