% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEDbaseR.R
\name{getMetadata}
\alias{getMetadata}
\title{Get metadata}
\usage{
getMetadata(id, record_type = c("bed", "bedset", "objects"))
}
\arguments{
\item{id}{BEDbase record identifier or object identifier}

\item{record_type}{bed, bedset, or objects}
}
\value{
a list with metadata
}
\description{
Get metadata
}
\examples{
getMetadata("421d2128e183424fcc6a74269bae7934")

}
