% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_obj_id}
\alias{make_obj_id}
\title{Construct object identifier}
\usage{
make_obj_id(rec_id, rec_type = "bed", result_id = "bedfile")
}
\arguments{
\item{rec_id}{character() BEDbase record identifier}

\item{rec_type}{character() (default bed) BEDbase record type}

\item{result_id}{character() (default bedfile) BEDbase result identifier}
}
\value{
String
}
\description{
Construct object identifier
}
\examples{
obj_id <- make_obj_id("bbad85f21962bb8d972444f7f9a3a932")
}
