% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{operations,BEDbase-method}
\alias{operations,BEDbase-method}
\title{Display API functions}
\usage{
\S4method{operations}{BEDbase}(x, ..., .deprecated = FALSE)
}
\arguments{
\item{x}{BEDbase() object}

\item{...}{other options}

\item{.deprecated}{(default \code{FALSE}) if deprecated}
}
\value{
list() API endpoints
}
\description{
Display functions defined through the
\href{https://api.bedbase.org}{bedhost API} and their corresponding parameters.
}
\examples{
bedbase <- BEDbase()
operations(bedbase)

}
