% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{operations,BEDbase-method}
\alias{operations,BEDbase-method}
\title{Display API}
\usage{
\S4method{operations}{BEDbase}(x, ..., .deprecated = FALSE)
}
\arguments{
\item{x}{BEDbase object}

\item{...}{other options}

\item{.deprecated}{(default FALSE) if deprecated}
}
\description{
Display API
}
\examples{
api <- BEDbase()
operations(api)

}
