% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{schemas,BEDbase-method}
\alias{schemas,BEDbase-method}
\title{Display bedhost API schemas}
\usage{
\S4method{schemas}{BEDbase}(x)
}
\arguments{
\item{x}{BEDbase() object}
}
\value{
list() API endpoints
}
\description{
Display bedhost API schemas
}
\examples{
bedbase <- BEDbase()
schemas(bedbase)

}
