% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{setCache,BEDbase-method}
\alias{setCache,BEDbase-method}
\title{Set cache along path}
\usage{
\S4method{setCache}{BEDbase}(x, cache_path, quietly = TRUE)
}
\arguments{
\item{x}{BEDbase() object}

\item{cache_path}{character(1)}

\item{quietly}{logical(1) (default \code{TRUE}) display messages}
}
\value{
\link[BiocFileCache:BiocFileCache-class]{BiocFileCache()} object of BED
files
}
\description{
Create a cache for BED files and BEDsets like
\href{https://docs.bedbase.org/geniml}{Geniml BBClient's cache}:
cache_path
bedfiles
a/f/afile.bed.gz
bedsets
a/s/aset.txt
}
\examples{
bedbase <- BEDbase(tempdir())
bedbase <- setCache(bedbase, "/tmp")

}
