% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{setCache,BEDbase-method}
\alias{setCache,BEDbase-method}
\title{Set cache path}
\usage{
\S4method{setCache}{BEDbase}(x, cache_path, quietly = TRUE)
}
\arguments{
\item{x}{BEDbase(1) object}

\item{cache_path}{character(1)}

\item{quietly}{logical(1) (default \code{TRUE}) display messages}
}
\value{
BiocFileCache(1) object of BED files
}
\description{
Set cache path
}
\examples{
bedbase <- BEDbase(tempdir())
bedbase <- setCache(bedbase, "/tmp")

}
