% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedbaser.R
\name{tags,BEDbase-method}
\alias{tags,BEDbase-method}
\title{Display tags}
\usage{
\S4method{tags}{BEDbase}(x, .tags, .deprecated = FALSE)
}
\arguments{
\item{x}{BEDbase(1) object}

\item{.tags}{character() tags for filtering operations}

\item{.deprecated}{(default \code{FALSE}) if deprecated}
}
\value{
list(1) API end points
}
\description{
Display tags
}
\examples{
bedbase <- BEDbase()
tags(bedbase)

}
